LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-scrolling-003.tentative-expected.txt



PASS Initial scroll position
PASS Scroll to 40,60
PASS Reattach at 40,60
FAIL Redisplay at 40,60 assert_equals: e1 x expected 120 but got 100
FAIL Scroll to 0,0 and relayout assert_equals: e1 x expected 120 but got 100
PASS Redisplay at 0,0
